Fix bugs in storage.py
It sometimes emits error transaction conflict.
If we choose to use explicit transaction mode, then it continues
to use them for all write transactions. But it didn't.
And it causes implicit transaction and stay in active.
Then it prohibits to start new transaction.
To prevent this, we uses explicit tranaction for all write operation.